+++ benches/bench_compare_l.rs
@@ -1,4 +1,6 @@
 use std::num::NonZeroU32;
+use std::thread::sleep;
+use std::time::Duration;
 
 use glassbench::*;
 use image::imageops;
@@ -46,7 +48,7 @@ pub fn bench_downscale_l(bench: &mut Bench) {
             let filter = match alg_name {
                 "Nearest" => {
                     // resizer doesn't support "nearest" algorithm
-                    task.iter(|| {});
+                    task.iter(|| sleep(Duration::new(0, 1)));
                     return;
                 }
                 "Bilinear" => resize::Type::Triangle,